Manual 2/requirement_1/MNIST Digit Recognition Project/results_SGClassifier/supplementary_1/code/requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
joblib==1.5.0
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.0.2
opencv-python==4.11.0.86
packaging==24.2.0
pandas==2.2.2
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
gdown
seaborn==0.13.2
